+++ AudioFX.cpp
@@ -11,7 +11,6 @@
 #include "constants.h"
 #include "utility.h"
 #include "AudioFX.h"
-#include "Volume3.h"
 
 #ifdef HARDWARE_SOUNDBOARD_ADAFRUIT
   bool SFX_ADAFRUIT = 1;
@@ -19,8 +18,6 @@
 #ifdef HARDWARE_SOUNDBOARD_JQ6500
   bool SFX_ADAFRUIT = 0;
 #endif
-
-Volume vol;
 
 AudioFX::AudioFX(Adafruit_Soundboard *sbref) {
   sfx = sbref;
@@ -32,11 +29,18 @@ AudioFX::AudioFX(JQ6500_Serial *jref) {
 
 void AudioFX::start_tone(int hz) {
   serial_print("Tone");
-  vol.tone(PIN_SPEAKER, TONE_DROP_HZ,TONE_VOLUME);
+  tone(PIN_SPEAKER, TONE_DROP_HZ);
 }
 
 void AudioFX::stop_tone() {
-  vol.noTone();
+  noTone(PIN_SPEAKER);
+}
+
+void AudioFX::play_power_on() {
+  tone(PIN_SPEAKER, 1000, 100);
+  delay(100);
+  tone(PIN_SPEAKER, 1333, 600);
+  delay(600);
 }
 
 void AudioFX::play_sound_samples() {
@@ -46,19 +50,17 @@ void AudioFX::play_sound_samples() {
   }
 
   serial_print("Play Sample1");
-  play_sample(SFX_PREP);
+  play_sample(SFX_PREP,2013);
   delay(1750);
   serial_print("Play Sample2");
-  play_sample(SFX_WATCH_GATE);
+  play_sample(SFX_WATCH_GATE,2330);
   serial_print("SFX Done");
 }
 
 void AudioFX::play_abort() {
-  vol.tone(PIN_SPEAKER, TONE_ABORT_1_HZ, TONE_VOLUME);
+  tone(PIN_SPEAKER, TONE_ABORT_1_HZ, DELAY_ABORT_TONE_1_MS);
   delay(DELAY_ABORT_TONE_1_MS);
-  vol.tone(PIN_SPEAKER, TONE_ABORT_2_HZ, TONE_VOLUME);
-  delay(DELAY_ABORT_TONE_2_MS);
-  vol.noTone();
+  tone(PIN_SPEAKER, TONE_ABORT_2_HZ, DELAY_ABORT_TONE_2_MS);
 }
 
 void AudioFX::play_sample(uint8_t track) {
@@ -72,26 +74,38 @@ void AudioFX::play_sample(uint8_t track) {
   }
   else
   {
+    // NOTE - for some annoying reason, serial transmission
+    // from the JQ6500 causes interference while audio is playing.
+    // We need to use the method below (passing in playtime) to
+    // finish playing when the track finishes. It is a crap way to do
+    // it, but for $3 for a JQ6500, I can live with it.
     serial_print("SFX playing on QJ6500 device");
     jfx->playFileByIndexNumber(track);
     delay(50);
-    while(jfx->getStatus() != MP3_STATUS_STOPPED)
+    while(jfx->getStatus() == MP3_STATUS_PLAYING)
     {
          // wait for sample to finish
+         // debug
+         //serial_print_val("SFX Status",jfx->getStatus());
     }
+    serial_print("SFX done");
   }
   return;
 }
 
-void AudioFX::play_coin_up() {
-  int start_vol = 130;
-  vol.tone(PIN_SPEAKER,1025,start_vol); // pa
-  delay(70);
-  int v = start_vol;
-  while(v > 0){
-    vol.tone(PIN_SPEAKER,2090,v); // ting!
-    delay(3);
-    v--;
+// needed for JQ6500, because cheap crap.
+void AudioFX::play_sample(uint8_t track, int playTime) {
+  if (SFX_ADAFRUIT) {
+    serial_print("SFX playing on Adafruit device");
+    sfx->playTrack(track);
+    delay(playTime);
   }
-  vol.noTone();
+  else
+  {
+    serial_print("SFX playing on QJ6500 device");
+    jfx->playFileByIndexNumber(track);
+    delay(playTime);
+  }
+  serial_print("SFX done");
+  return;
 }